Small bugfixes

+++ gwt-d3-api/src/main/java/com/github/gwtd3/api/layout/Node.java
@@ -54,7 +54,7 @@ public class Node
      * @param name of attribute as a string
      * @param value as a double
      */
-    public final native int setAttr(String name, double value) /*-{
+    public final native double setAttr(String name, double value) /*-{
 		return this[name] = value;
     }-*/;
